+++ noodles-cram/src/codecs/aac.rs
@@ -1,9 +1,10 @@
+//! Adaptive arithmetic coder.
+
 mod decode;
 mod encode;
 mod flags;
 mod model;
 mod range_coder;
 
-pub use self::{
-    decode::decode, encode::encode, flags::Flags, model::Model, range_coder::RangeCoder,
-};
+pub use self::flags::Flags;
+pub(crate) use self::{decode::decode, encode::encode, model::Model, range_coder::RangeCoder};