+++ noodles-cram/src/codecs/aac/flags.rs
@@ -1,12 +1,21 @@
 bitflags::bitflags! {
+    /// Adaptive arithmetic coder flags.
     pub struct Flags: u8 {
+        /// Order-1 entropy coding.
         const ORDER = 0x01;
+        /// Reserved.
         const RESERVED = 0x02;
+        /// Use an external codec (bzip2).
         const EXT = 0x04;
+        /// Interleave 32 rANS states.
         const STRIPE = 0x08;
+        /// Discard the uncompressed data size.
         const NO_SIZE = 0x10;
+        /// Data is uncompressed.
         const CAT = 0x20;
+        /// Use run-length encoding.
         const RLE = 0x40;
+        /// Use bit packing.
         const PACK = 0x80;
     }
 }