+++ noodles-cram/CHANGELOG.md
@@ -12,7 +12,7 @@
 
   * cram/codecs: Add fqzcomp encoder.
 
-  * cram/codecs: Increase visibilities of `rans_4x8::Order`,
+  * cram/codecs: Increase visibilities of `aac::Flags`, `rans_4x8::Order`, and
     `rans_nx16::Flags`.
 
   * cram/codecs/name_tokenizer/encode: Add `DIGITS0`, `DIGITS`, `DELTA`,